Add rabbit connection with msg handling

+++ internal/model/presence.go
@@ -8,6 +8,6 @@ type Presence struct {
 	ID int
 	MAC string
 	LastDetected *time.Time
-	RoomID int64
+	RoomID int
 	Active bool
 }
